Add week entity, add relationship to issue entity
One issue might be covered in multiple weeks as well as a single week
will cover multiple issues.

The total workload might still be something that needs some tweaking, as
currently it'll look like the workload per week is too large if a large
issue (workload > maximum) is assigned to a week and developer.